navbar: Move "Code of Conduct" to the end
This "how do I interact with this project" meta information is similar
to the license, and the current README links to both the license and
code of conduct at the end of the "Contributing" section.  With this
shift, the information we're explicitly teaching is on the left of the
navbar, and as you go right the navbar entries become more meta.